@@ -1,53 +1,57 @@
 {
-  "name": "starter",
-  "version": "0.0.0",
+  "name": "essentials",
+  "version": "3.0.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve -o",
+    "start": "ng serve",
+    "proxy": "ng serve --proxy-config proxy.conf.json",
     "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "tsc": "tsc"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "11.0.2",
-    "@angular/common": "11.0.2",
-    "@angular/compiler": "11.0.2",
-    "@angular/core": "11.0.2",
-    "@angular/forms": "11.0.2",
-    "@angular/platform-browser": "11.0.2",
-    "@angular/platform-browser-dynamic": "11.0.2",
-    "@angular/router": "11.0.2",
-    "angular-oauth2-oidc": "^10.0.3",
-    "bootstrap": "^4.0.0-beta.2",
-    "core-js": "^2.5.1",
-    "rxjs": "^6.5.4",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "@angular/animations": "15.2.9",
+    "@angular/common": "15.2.9",
+    "@angular/compiler": "15.2.9",
+    "@angular/core": "15.2.9",
+    "@angular/forms": "15.2.9",
+    "@angular/platform-browser": "15.2.9",
+    "@angular/platform-browser-dynamic": "15.2.9",
+    "@angular/router": "15.2.9",
+    "angular-oauth2-oidc": "^15.0.1",
+    "bootstrap": "^4.6.2",
+    "rxjs": "^7.8.0",
+    "tslib": "~2.3.1",
+    "zone.js": "~0.12.0"
   },
   "devDependencies": {
-    "@angular/cli": "11.0.2",
-    "@angular/compiler-cli": "11.0.2",
-    "@angular/language-service": "11.0.2",
-    "@types/jasmine": "~3.6.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.1.1",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~3.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "4.0.5",
-    "@angular-devkit/build-angular": "~0.1100.2"
+    "@angular-builders/jest": "^15.0.0",
+    "@angular-devkit/build-angular": "15.2.8",
+    "@angular-eslint/builder": "^15.2.1",
+    "@angular-eslint/eslint-plugin": "^15.2.1",
+    "@angular-eslint/eslint-plugin-template": "^15.2.1",
+    "@angular-eslint/template-parser": "^15.2.1",
+    "@angular/cli": "~15.2.8",
+    "@angular/compiler-cli": "15.2.9",
+    "@angular/language-service": "15.2.9",
+    "@briebug/jest-schematic": "^6.0.0",
+    "@cypress/schematic": "^2.5.0",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^16.18.12",
+    "@typescript-eslint/eslint-plugin": "^5.51.0",
+    "@typescript-eslint/parser": "^5.51.0",
+    "cypress": "^11.2.0",
+    "eslint": "^8.34.0",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.4.2",
+    "prettier": "^2.8.4",
+    "typescript": "~4.8.4",
+    "yarn-upgrade-all": "^0.7.2"
   }
 }
