@@ -1,24 +1,21 @@
-import { Component, OnInit, ViewChild, ElementRef, Renderer2 } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
-import { Location, LocationStrategy, PathLocationStrategy } from '@angular/common';
+import { Component } from '@angular/core';
 
 @Component({
-    selector: 'navbar-cmp',
-    templateUrl: 'navbar.component.html'
+  selector: 'app-navbar',
+  templateUrl: 'navbar.component.html'
 })
 export class NavbarComponent {
-    
-    private sidebarVisible: boolean = false;
+  private sidebarVisible = false;
 
-    sidebarToggle(){
-        var body = document.getElementsByTagName('body')[0];
+  sidebarToggle(): void {
+    const body = document.getElementsByTagName('body')[0];
 
-        if(this.sidebarVisible == false){
-            body.classList.add('nav-open');
-            this.sidebarVisible = true;
-        } else {
-            this.sidebarVisible = false;
-            body.classList.remove('nav-open');
-        }
+    if (this.sidebarVisible == false) {
+      body.classList.add('nav-open');
+      this.sidebarVisible = true;
+    } else {
+      this.sidebarVisible = false;
+      body.classList.remove('nav-open');
     }
+  }
 }
